Feat: LadderCreator 인터페이스로 분리, 이를 구현한 임시 클래스인 StaticLadderCreator생성 - 리팩토링 필요
