+++ src/main/java/ladder/LadderRunner.java
@@ -14,4 +14,10 @@ public class LadderRunner {
         }
         return position.getValue();
     }
+
+    public void PrintLadder() {
+        for (int i = 1; i < rows.length+1; i++) {
+            System.out.println(i + "번째 사람의 결과 -> " + run(Position.of(i)));
+        }
+    }
 }
